% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgnUtilities.R
\name{add_colors_to_sbgnml}
\alias{add_colors_to_sbgnml}
\title{Add colors to SBGNML}
\usage{
add_colors_to_sbgnml(sbgnml_file, overall_df,
  color_xml_file = system.file("templates/color_template.xml", package =
  "sbgnvizShiny"))
}
\arguments{
\item{sbgnml_file}{an SBGNML filename}

\item{overall_df}{see extract_elements}

\item{color_xml_file}{a template}
}
\value{
XML content (from xml2) with added color information
}
\description{
Add colors to SBGNML
}
