% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgnUtilities.R
\name{extract_elements}
\alias{extract_elements}
\title{Extract information for SBGNML elements from an SBGN file}
\usage{
extract_elements(sbgnml_file, stroke_color = "#555555",
  compartment_process_fill_color = "#ffffff")
}
\arguments{
\item{sbgnml_file}{an SBGNML filename}

\item{a}{default stroke color (e.g., #555555)}

\item{a}{default compartment and process fill color (e.g., #ffffff)}
}
\value{
a data.frame with information about the extracted elements: id, label, type, stroke_color, fill_color
}
\description{
Extract information for SBGNML elements from an SBGN file
}
