% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgnUtilities.R
\name{get_color_xml}
\alias{get_color_xml}
\title{Get color XML content}
\usage{
get_color_xml(color_id, color_value)
}
\arguments{
\item{color_id}{a string id (e.g., color_1)}

\item{color_value}{a string hex color (e.g., #ffffff)}
}
\value{
a string XML element
}
\description{
Get color XML content
}
