% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgnUtilities.R
\name{get_style_xml}
\alias{get_style_xml}
\title{Get style XML content}
\usage{
get_style_xml(style_id, id_list, type, stroke_color, fill_color = NULL,
  font_size = 12, font_family = "Helvetica", font_weight = "normal",
  font_style = "normal", stroke_width = 2)
}
\arguments{
\item{style_id}{a string id (e.g., color_1)}

\item{id_list}{a vector of SBGNML IDs}

\item{stroke_color}{a color ID (e.g., color_1)}

\item{fill_color}{a color ID (e.g., color_1)}

\item{font_size}{a numeric font size}

\item{font_family}{a font family name (should be compatible with web browsers)}

\item{font_weight}{a CSS font weight}

\item{stroke_width}{a numeric stroke width}
}
\value{
a string XML element
}
\description{
Get style XML content
}
