% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgnUtilities.R
\name{map_colors}
\alias{map_colors}
\title{Map colors with queries}
\usage{
map_colors(query_colors, overall_df, default_fill_color = "#ffffff")
}
\arguments{
\item{query_colors}{a two-column data.frame with query and fill_color columns. The query will be grepped against the labels in overall_df.}

\item{overall_df}{see extract_elements function}

\item{default_fill_color}{a default color (e.g., #ffffff) mapped if not matched by any query}
}
\value{
overall_df (see extract_elemnts function with updated colors
}
\description{
Map colors with queries
}
