% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgnvizShiny.R
\name{renderSbgnvizShiny}
\alias{renderSbgnvizShiny}
\title{More shiny plumbing -  a sbgnvizShiny wrapper for htmlwidget standard rendering operation}
\usage{
renderSbgnvizShiny(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates an HTML widget.}

\item{env}{environment in which to evaluate expr.}

\item{quoted}{logical specifies whether expr is quoted ("useuful if you want to save an expression in a variable").}
}
\value{
not sure
}
\description{
More shiny plumbing -  a sbgnvizShiny wrapper for htmlwidget standard rendering operation
}
