% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbgnvizShiny.R
\name{sbgnvizShinyOutput}
\alias{sbgnvizShinyOutput}
\title{Standard shiny ui rendering construct}
\usage{
sbgnvizShinyOutput(outputId, width = "100\%", height = "400")
}
\arguments{
\item{outputId}{the name of the DOM element to create.}

\item{width}{integer  optional initial width of the widget.}

\item{height}{integer optional initial height of the widget.}
}
\value{
a reference to an htmlwidget
}
\description{
Standard shiny ui rendering construct
}
\examples{
\dontrun{
  mainPanel(sbgnvizShinyOutput('sbgnvizShiny'), width=10)
}

}
